Refactored test configuration
