+++ web/src/test/java/org/cbioportal/weblegacy/ApiControllerConfig.java
@@ -31,12 +31,6 @@
 */
 package org.cbioportal.weblegacy;
 
-import org.cbioportal.persistence.CancerTypeRepository;
-import org.cbioportal.persistence.MolecularProfileRepository;
-import org.cbioportal.persistence.mybatis.util.CacheMapUtil;
-import org.cbioportal.persistence.PatientRepository;
-import org.cbioportal.persistence.SampleListRepository;
-import org.cbioportal.persistence.StudyRepository;
 import org.cbioportal.service.CosmicCountService;
 import org.mockito.Mockito;
 import org.mskcc.cbio.portal.persistence.CancerTypeMapperLegacy;
@@ -119,29 +113,4 @@ public class ApiControllerConfig extends WebMvcConfigurerAdapter {
     public StudyMapperLegacy studyMapper() {
         return Mockito.mock(StudyMapperLegacy.class);
     }
-    @Bean
-    public CacheMapUtil cacheMapUtil() {
-        return Mockito.mock(CacheMapUtil.class);
-    }
-    @Bean
-    public PatientRepository patientRepository() {
-        return Mockito.mock(PatientRepository.class);
-    }
-    @Bean
-    public CancerTypeRepository cancerTypeRepository() {
-        return Mockito.mock(CancerTypeRepository.class);
-    }
-    @Bean
-    public StudyRepository studyRepository() {
-        return Mockito.mock(StudyRepository.class);
-    }
-    @Bean
-    public MolecularProfileRepository molecularProfileRepository() {
-        return Mockito.mock(MolecularProfileRepository.class);
-    }
-    @Bean
-    public SampleListRepository sampleListRepository() {
-        return Mockito.mock(SampleListRepository.class);
-    }
-
 }
