+++ web/src/test/java/org/cbioportal/weblegacy/MutationalSignatureControllerTest.java
@@ -31,6 +31,7 @@
 */
 package org.cbioportal.weblegacy;
 
+import org.cbioportal.web.config.CacheMapUtilConfig;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -55,7 +56,7 @@ import org.springframework.web.context.WebApplicationContext;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @WebAppConfiguration
-@ContextConfiguration(classes = {MutationalSignatureControllerConfig.class, CustomObjectMapper.class})
+@ContextConfiguration(classes = {MutationalSignatureControllerConfig.class, CustomObjectMapper.class, CacheMapUtilConfig.class})
 public class MutationalSignatureControllerTest {
     @Autowired
     private WebApplicationContext webApplicationContext;
