+++ web/src/test/java/org/cbioportal/weblegacy/StructuralVariantControllerTest.java
@@ -32,6 +32,7 @@
 
 package org.cbioportal.weblegacy;
 
+import org.cbioportal.web.config.CacheMapUtilConfig;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -62,7 +63,7 @@ import org.springframework.web.context.WebApplicationContext;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @WebAppConfiguration
-@ContextConfiguration(classes = {StructuralVariantControllerTestConfig.class, CustomObjectMapper.class})
+@ContextConfiguration(classes = {StructuralVariantControllerTestConfig.class, CustomObjectMapper.class, CacheMapUtilConfig.class})
 public class StructuralVariantControllerTest {
     @Autowired
     private StructuralVariantService structuralVariantServiceMock;
