+++ web/src/test/java/org/cbioportal/weblegacy/StructuralVariantControllerTestConfig.java
@@ -32,12 +32,6 @@
 
 package org.cbioportal.weblegacy;
 
-import org.cbioportal.persistence.CancerTypeRepository;
-import org.cbioportal.persistence.MolecularProfileRepository;
-import org.cbioportal.persistence.mybatis.util.CacheMapUtil;
-import org.cbioportal.persistence.PatientRepository;
-import org.cbioportal.persistence.SampleListRepository;
-import org.cbioportal.persistence.StudyRepository;
 import org.mockito.Mockito;
 import org.mskcc.cbio.portal.service.StructuralVariantService;
 import org.springframework.context.annotation.Bean;
@@ -54,29 +48,5 @@ public class StructuralVariantControllerTestConfig extends WebMvcConfigurerAdapt
     @Bean
     public StructuralVariantService svService(){
         return Mockito.mock(StructuralVariantService.class);
-    }
-    @Bean
-    public CacheMapUtil cacheMapUtil() {
-        return Mockito.mock(CacheMapUtil.class);
-    }
-    @Bean
-    public PatientRepository patientRepository() {
-        return Mockito.mock(PatientRepository.class);
-    }
-    @Bean
-    public CancerTypeRepository cancerTypeRepository() {
-        return Mockito.mock(CancerTypeRepository.class);
-    }
-    @Bean
-    public StudyRepository studyRepository() {
-        return Mockito.mock(StudyRepository.class);
-    }
-    @Bean
-    public MolecularProfileRepository molecularProfileRepository() {
-        return Mockito.mock(MolecularProfileRepository.class);
-    }
-    @Bean
-    public SampleListRepository sampleListRepository() {
-        return Mockito.mock(SampleListRepository.class);
     }
 }
