+++ web/src/test/resources/applicationContext-web-test.xml
@@ -11,25 +11,8 @@
      http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc.xsd
      http://www.springframework.org/schema/aop http://www.springframework.org/schema/aop/spring-aop.xsd">
 
-    <!-- define beans needed for testing -->
-    <bean id="cacheMapUtil" class="org.mockito.Mockito" factory-method="mock"> 
-        <constructor-arg value="org.cbioportal.persistence.mybatis.util.CacheMapUtil" />
-    </bean> 
-    <bean id="patientRepository" class="org.mockito.Mockito" factory-method="mock"> 
-        <constructor-arg value="org.cbioportal.persistence.PatientRepository" />
-    </bean> 
-    <bean id="cancerTypeRepository" class="org.mockito.Mockito" factory-method="mock"> 
-        <constructor-arg value="org.cbioportal.persistence.CancerTypeRepository" />
-    </bean> 
-    <bean id="studyRepository" class="org.mockito.Mockito" factory-method="mock"> 
-        <constructor-arg value="org.cbioportal.persistence.StudyRepository" />
-    </bean> 
-    <bean id="molecularProfileRepository" class="org.mockito.Mockito" factory-method="mock"> 
-        <constructor-arg value="org.cbioportal.persistence.MolecularProfileRepository" />
-    </bean> 
-    <bean id="sampleListRepository" class="org.mockito.Mockito" factory-method="mock"> 
-        <constructor-arg value="org.cbioportal.persistence.SampleListRepository" />
-    </bean> 
+    <!-- load beans from shared test configuration -->
+    <bean class="org.cbioportal.web.config.CacheMapUtilConfig"/>
 
     <!-- load the beans from the prodution configuration xml file -->
     <import resource="classpath:applicationContext-web.xml" />
